Two-way synchronization between sqlite and indexedDB

+++ server/api/syncTodos.ts
@@ -1,16 +1,16 @@
 import db from './db';
 
 type Todo = {
- id: number | null,
- idbId: number,
- content: string,
- done: number,
- deleted: boolean,
- synced: boolean
+  id: number | null,
+  idbId: number,
+  content: string,
+  done: number,
+  deleted: boolean,
+  synced: boolean
 }
 
 export default defineEventHandler(async (event) => {
-const todos = await readBody(event);
+  const todos = await readBody(event);
 
   if (!todos?.length) {
     return {
@@ -20,6 +20,8 @@ const todos = await readBody(event);
 
   return new Promise(async (resolve, reject) => {
     try {
+      const syncedIds: Number[] = [];
+
       const updatedRows = await Promise.all(todos.map(async (todo: Todo) => {
         const {
           id,
@@ -28,9 +30,11 @@ const todos = await readBody(event);
           deleted
         } = todo;
 
-        if (deleted) {
+        if (deleted && id) {
           // Delete the row from the database
           await db.run('DELETE FROM todos WHERE id = ?;', id);
+
+          syncedIds.push(id);
 
           return {
             ...todo,
@@ -41,6 +45,8 @@ const todos = await readBody(event);
           // Create a new row in the database
           const result = await db.query!('INSERT INTO todos (content, done) VALUES (?, ?) RETURNING id', [content, done]);
 
+          syncedIds.push(result.rows[0].id);
+
           return {
             ...todo,
             id: result.rows[0].id,
@@ -49,17 +55,31 @@ const todos = await readBody(event);
         } else {
           // Update the row in the database
           await db.run('UPDATE todos SET content = ?, done = ? WHERE id = ?;', content, done, id);
+
+          syncedIds.push(id);
         }
-  
+
         return {
           ...todo,
           synced: true
         };
       }));
-  
+
+      const syncedIdsStringified = syncedIds.join(',');
+
+      let otherRows: Todo[] = await new Promise((resolve, reject) => {
+        db.all(`SELECT * FROM todos WHERE ID NOT IN (${syncedIdsStringified})`, (error, rows: Todo[]) => {
+          if (error) {
+            reject(error);
+          } else {
+            resolve(rows);
+          }
+        });
+      });
+
       resolve({
         status: 'SUCCESS',
-        updatedRows
+        rows: [...updatedRows, ...otherRows],
       });
     } catch (error) {
       console.error('Error updating database:', error);
